Fixed metrics reporting and cache clearing logic.

+++ PdtCfwComponents/LabBenchStudios/Src/Main/ProgrammingDigitalTwins/Common/IDataHistorianCache.cs
@@ -51,7 +51,7 @@ namespace LabBenchStudios.Pdt.Common
         /// 
         /// </summary>
         /// <returns></returns>
-        public long GetCacheMemoryUsage();
+        public double GetCacheMemoryUsage();
 
         /// <summary>
         ///
